Well,
at this point, I've decided to patch the SYSCALL handler, replacing the address of the sys_call_table by my own.
Atm, it's not replaced. But everything is working on this kernel v4.13 of ubuntu.
I'm commiting because I'm publishing on twitter the news.

The idea is to own the real sys_call_table and our own which is implanted on the system.
Then we can call the reals and the programs are going to call the ones we want.
Then when the anti malware tool searches on the real sct, would find everything normal.
It must search the sys_call_table from the opcodes inside the SYSCALL handler to know the hooks.
Or debug the kernel in runtime, breaking point in SYSCALL handler.

The int 0x80 32 bit compatibility mode, is going to be replaced also. It's called ia32_sys_call_table, but now I haven't implemented its search.

Also, I located the internal function for the wrappers set_memory_rw and _ro because we need to replace the sys_call_table address inside the handlers.

In the both handlers, (fastcall and slow_path) the address storing the sys_call_table address are located.

Have a nice coding! =)