Yeah!
I was able to crash the system by replacing the sys_call_table address inside the SYSCALL handlers, by an empty one.

Thanks to http://vulnfactory.org/blog/2011/08/12/wp-safe-or-not/ the information on this website, I was able to develop disable_wp() and enable_wp() routines.

Now, let's begin the tests with int $0x80, so let's try to find the handler and see what's inside, because I saw another symbol for sys_call_table and 32bits, but I'm not sure if it's for native 32bit kernel or for compat support.

In theory, now if hooking syscalls from the new sys_call_table, only would take effect on SYSCALL calls. Not in int $0x80.