I commit now that I found the way of reading request shell by hooking sys_read() by the use of MSG_PEEK flag of sys_recvmg() and later I will delete the before implementation, which is untested and hard-way before knowing about MSG_PEEK.
I have to clean the kernel.c and atm linked-lists are not used...

Also disass_search_* functions are unused because I found another way.

Well, I think that the way of this generic rootkit, is not accesing the kernel structures and yes to use syscalls before returning and analyzing data. Filtering is done in my_read64 atm.
That way is the right for a 4.x generic rk.